Uploaded computer Timer program
This program is to log how long your computer has run in total.
it is meant to be put into the start menu (with a launcher) so that it
runs when your computer starts up. The program saves the current time
every 5 minutes, and starts from where it left off when your computer
starts up.